+++ src/time.c
@@ -17,6 +17,7 @@
 
 #include "time.h"
 #include "parodus_log.h"
+#include <errno.h>
 
 /*----------------------------------------------------------------------------*/
 /*                             External Functions                             */
@@ -24,7 +25,10 @@
 
 void getCurrentTime(struct timespec *timer)
 {
-	clock_gettime(CLOCK_REALTIME, timer);
+	if( clock_gettime(CLOCK_REALTIME, timer) == -1 )
+	{
+		ParodusError("clock gettime returns errno %d\n", errno );
+	}
 }
 
 uint64_t getCurrentTimeInMicroSeconds(struct timespec *timer)
@@ -34,7 +38,7 @@ uint64_t getCurrentTimeInMicroSeconds(struct timespec *timer)
     ParodusPrint("timer->tv_sec : %lu\n",timer->tv_sec);
     ParodusPrint("timer->tv_nsec : %lu\n",timer->tv_nsec);
     systime = (uint64_t)timer->tv_sec * 1000000L + timer->tv_nsec/ 1000;
-    return systime;	
+    return systime;
 }
 
 long timeValDiff(struct timespec *starttime, struct timespec *finishtime)