+++ src/upstream.h
@@ -47,11 +47,15 @@ typedef struct UpStreamMsg__
 void packMetaData();
 void *handle_upstream();
 void *processUpstreamMessage();
+void registerRBUSlistener();
 int getDeviceId(char **device_id, size_t *device_id_len);
-void sendUpstreamMsgToServer(void **resp_bytes, size_t resp_size);
+int sendUpstreamMsgToServer(void **resp_bytes, size_t resp_size);
 void getServiceNameAndSendResponse(wrp_msg_t *msg, void **msg_bytes, size_t msg_size);
 void createUpstreamRetrieveMsg(wrp_msg_t *message, wrp_msg_t **retrieve_msg);
 void set_global_UpStreamMsgQ(UpStreamMsg * UpStreamQ);
+#ifdef WAN_FAILOVER_SUPPORTED
+int subscribeCurrentActiveInterfaceEvent();
+#endif
 UpStreamMsg * get_global_UpStreamMsgQ(void);
 pthread_cond_t *get_global_nano_con(void);
 pthread_mutex_t *get_global_nano_mut(void);