+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Add additional HTTP headers for call to Themis from Convey
+- Add callback handler for initial cloud connection status change event
+- Fix Parodus connection stuck on interface up down received together
+- Update to use nopoll version 1.0.3
 
 ## [v1.1.5]
 - Add additional HTTP headers for call to Themis from Convey